Add a sensitivity keyword in the Makefile to make RMG-Py with DASPK.
Because try and except statements don't work properly in the cythonized
python code, I am relying on a settings.pxi file
to check whether or not to compile the cython code with DASPK or DASSL,
based on whether the user desires it.  In order to compile with DASPK,
the user must have pydaspk module installed on their machine and use the
command

make sensitivity=on

This then checks whether pydaspk can be imported and also writes the
settings.pxi to reflect that DASPK=1 is set.  This is probably just
one of many ways to conditionally import in cython.