moved degeneracy algorithm to react.py
moved the degeneracy algorithm to react.py to allow for various
resonance isomers' reactions to be combined and incremented in
the degeneracy algorithm. This needed to be done before `deflate`,
so a new method `reactSpecies` was created that finds degeneracy
before deflating. This slightly alters the parallized algorithm with
species being parallelized instead of molecules.