reduced degeneracy for identical reactants
reactants which are identical should have half the reaction rate
when using the degeneracy method, since the translational component
of their rates is already taken into account in their concentration,
so them swapping places doesn't have a real effect on reaction
rate.

This was placed in a separate method in the react module to be applied
separately from finding degeneracies. The calculateDegeneracy function
was modified to allow for degeneracy to be applied separately from
the reduction of same reactant degeneracy. This method works on the
rxn.reverse attribute as well if it comes from a family with
ownReverse=True