check for family and templates in findDegeneracies
modified the findDegeneracies to fix bug that resulted from
finding the same reactions if forward and backward directions are
flipped (like in beta scission reactions). At the same time
this commit requires the same family for the reaction to go through
adding a requirement for the same template can be done later after
multiple transition states are supported

To fix a resonance structure bug, the method will now
use Species objects with resonance structures in find degneracy
to allow RMG to find resonance isomers. Removes the species for
deflating