modified calculateDegeneracy using resonance isomers and templates
altered the code in calculateDegeneracy to find the resonance
isomers of the molecule and combine them to obtain a more
realistic degeneracy number. This ensures that reactions found
in the reverse direction have the same degeneracy as when found
in the forward direction (when the family ownReverse isn't true)

also used the reaction template to determine the degeneracy of which
transition state to return.