Regression(?) test for parsing [CH] methylidyne SMILES
For now, this is a regression test, to ensure we make what
we currently make, which is the quartet with 3 unpaired electrons.
However, Wikipedia reports:
  The ground state is a doublet radical with one unpaired electron,
  and the first two excited states are a quartet radical with three
  unpaired electrons and a doublet radical with one unpaired electron.
  With the quartet radical only 71 kJ above the ground state, a sample
  of methylidyne exists as a mixture of electronic states even at
  room temperature, giving rise to complex reactions.

So perhaps we should make the doublet radical with one unpaired electron?
(but how would it react as such? and what do we have in the database?)