allow comparison of identical reactions by rxn.isIsomorphic
created an option for using the rxn.isIsomorphic method to also look at
atom labels. At the same time, the method was refactored to avoid code
duplication by creating a static method which compares lists of species' isomorphism.
this method could also be used in other methods in different parts of RMG-Py