Add more information in supporting_information.csv
This commit refactors where the supporting information is saved
to be at the end of StatMechJob.load method instead of partially
in the __init__ method and throughout the load method.

This commit also adds the symmetry, optical isomers,
energies (with and without zero-point energy), T1 diagnostic,
and geometry to the exported csv file.

This commit also prevents from creating the supporting_information.csv
file when no species will be output, simplifying the process for
other users of Arkane.

Try/except statements are added to reduce potential for errors.