Refactor plotHinderedRotor
Previously, plotHinderedRotor was called under the `load` method,
which gave limited flexibility as to where the hindered rotor plot
would be saved and was also unintuitive since it was not in a
method labeled `output`.

This commit refactors the plotHinderedRotor method by separating
it into two separate methods, one for internally storing the data,
and a second for writing the plot to disk. This should reduce
any potential errors in writing from impacting the loading and
creating of a model.

The refactor also saves the plots in the Arkane output folder,
which should be more intuitive for users.