'onlyHeterocyclics' option for mlEstimator added
A new option 'onlyHeterocyclics' is added for Machine Learning Thermo
Estimator. If 'onlyHeterocyclics' is True, the thermo of only heterocyclics
species are estimated by mlEstimator regardless of 'onlyCyclics'
setting. If 'onlyCyclics' is False and 'onlyHeterocyclics' is True, then
RMG will log a warning that 'onlyCyclics' should be also True since
heterocyclics is a subset of cyclics.
I also added lines that will log a warning when 'onlyCyclics' is False
and 'minCycleOverlap' is greater than zero since the machine learning
estimator is restricted to only cyclic species when 'minCycleOverlap' is
greater than zero.
If 'onlyCyclics' is False and 'heteroCyclics' is True and
'minCycleOverlap' is greater than zero, it will log a warning and the
machine learning estimator will be restricted to only heterocyclic
species with the speficied minimum cycle overlap.