Set Tlist when creating KineticsJob
Previously, there were two spots where a list of temperatures
was created for fitting high pressure kinetics in Arkane.
One fit using inverse temperature spacing and the other one fit
with linear temperature spacing. See #1671

This commit moves the location for creating a listing of temperatures
to the __init__ method, so there is only one spot where temperatures
are initialized. The default min and max in `generateKinetics` was
moved into the __init__ method. Three unittests were written for
obtaining the proper temperatures in the __init__ method.

`generateKinetics` previously accepted an arguement Tlist, which
is already an attribute of KineticsJob, creating redundancy.
The Tlist argument was removed and self.Tlist was used
in this method instead, simplifying the code.

Calls to quantity.Quantity in the __init__ method were unnecessary
before setting them with quantity.Temperature, so the __init__
code was simplified by setting attributes with (value, 'K') instead.

Documentation was updated to better describe the three methods
of specifying temperature when calling `kinetics` in Arkane