Use correct energy and output debug message about energy origin
Previously the gaussian log file would read the SCF energy
if coupled cluster or similar methods were being used. This would
increase the error in calculation.

This commit allows for the correct energy to be obtained from MP2,
CCSD, and CCSD(T) methods. It also adds a debug message which
specifies what type of energy is being taken from the log file.