Fix reservoir state to work with tunneling
Though it wasn't documented, the reservoir state would lead to
substantial inaccuracies when applied to mechanisms with tunneling.
This was caused by the cutoff for reservoir height being set to
when all the path reactions out of the isomer are zero, which if
there is tunneling, doesn't happen (it just gets really really small).

This commit fixes that by setting the reservoir cutoff to be whenever
the rate is less than 1% of the rate above the transition state barrier.
For intramolecular hydrogen abstractions, this appears to make the
reservoir state cutoff about 8 kJ/mol lower than the barrier height.

This does not change the behavior when there is no tunneling but allows
the code to work with tunneling.